change map logic
A marker on a map can be on multiple routes.
Therefore only one marker should be created on the map with all routes without the link.
An each map where the routs are seperated the marker will be linked.
To link to different routes every place will have his own id -> name of the place + route name.